requirements.txt
cffi==1.16.0
cryptography==42.0.1
linecache2==1.0.0
pybase64==1.3.2
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
traceback2==1.4.0
unittest2==1.1.0
